元画像サイズ、コンポーネント整理
