--- src/app/_components/Keyvisual/style.module.scss
@@ -1,303 +1,0 @@
-@use '../../../styles/functions' as *;
-@use '../../../styles/variables' as *;
-@use '../../../styles/mixins' as mixins;
-
-/* ==========================================================================
-   1. メインジュアル
-   ========================================================================== */
-
-.keyvisual-section {
-  width: 100%;
-  margin: 0 auto;
-  font-family: 'Noto Sans JP', sans-serif;
-  color: $color-black;
-
-  .container {
-    display: flex;
-    justify-content: flex-end;
-    min-height: sp-vw(270);
-    background-image: url('/icon/title.jpg');
-    background-position: 80% center;
-    background-size: 125%;
-  }
-
-  .group {
-    text-align: center;
-    @include mixins.flex-column-center;
-  }
-
-  .keyvisual-img {
-    width: sp-vw(150);
-    height: sp-vw(80);
-    margin: sp-vw(5) 0;
-  }
-
-  .main-title {
-    margin: sp-vw(8) 0;
-    font-family: 'DM Sans', sans-serif;
-    font-size: sp-vw(17);
-    font-weight: bold;
-    color: $color-white;
-    text-shadow: 0 2px 8px rgb(0 0 0 / 20%);
-  }
-
-  .job-count {
-    display: inline-block;
-    padding: sp-vw(3) sp-vw(15);
-    margin: sp-vw(5) 0;
-    font-size: sp-vw(8);
-    color: $color-light-green;
-    background-color: $color-white;
-    border: 2px solid $color-light-green;
-    border-radius: $radius-sm;
-    box-shadow: 0 4px 10px rgb(0 0 0 / 8%);
-
-    .count-num {
-      font-size: sp-vw(20);
-      font-weight: bold;
-    }
-  }
-}
-
-/* ==========================================================================
-   2. 仕事検索フォーム
-   ========================================================================== */
-
-.search-form {
-  position: relative;
-  max-width: sp-vw($inner-sp-max-width);
-  padding: sp-vw(20) sp-vw(15) sp-vw(15);
-  margin: sp-vw(20) auto;
-  text-align: center;
-  background-color: $color-dark-blue;
-  border-radius: sp-vw(12);
-  box-shadow: 0 6px 16px rgb(0 0 0 / 6%);
-
-  .search-title {
-    font-size: sp-vw(22);
-    font-weight: bold;
-  }
-
-  .description {
-    margin: sp-vw(15) auto;
-    font-size: sp-vw(15);
-    color: $color-black;
-  }
-
-  .search-conditions {
-    display: flex;
-    gap: sp-vw(7);
-    align-items: center;
-    justify-content: center;
-  }
-
-  .group-action {
-    display: flex;
-    flex-direction: column;
-    gap: sp-vw(12);
-    align-items: center;
-  }
-
-  .select {
-    width: sp-vw(105);
-    padding: sp-vw(6);
-    padding-right: sp-vw(25);
-    margin: sp-vw(5) 0;
-    font-family: 'Noto Sans JP', sans-serif;
-    font-size: sp-vw(13);
-    color: $color-black;
-    text-align: left;
-    appearance: none;
-    background-color: $color-white;
-    background-image:
-      linear-gradient(45deg, transparent 50%, $color-primary-pink 50%),
-      linear-gradient(135deg, $color-primary-pink 50%, transparent 50%);
-    background-repeat: no-repeat;
-    background-position:
-      right sp-vw(15) center,
-      right sp-vw(8) center;
-    background-size: sp-vw(7) sp-vw(7);
-    border: 1.5px solid $color-primary-pink;
-    border-radius: $radius-sm;
-  }
-
-  .search-input {
-    width: sp-vw(320);
-    padding: sp-vw(4);
-    margin: sp-vw(8) 0;
-    font-family: 'Noto Sans JP', sans-serif;
-    font-size: sp-vw(14);
-    color: $color-text-sub;
-    text-align: left;
-    background-color: $color-white;
-    border: 1.5px solid $color-primary-pink;
-    border-radius: $radius-sm;
-  }
-
-  .search-button {
-    display: flex;
-    gap: sp-vw(20);
-    align-items: center;
-    width: sp-vw(160);
-    padding: sp-vw(10);
-    font-family: 'Noto Sans JP', sans-serif;
-    font-size: sp-vw(18);
-    font-weight: bold;
-    box-shadow: $shadow-pink;
-    @include mixins.pink-btn;
-  }
-
-  .search-img {
-    width: sp-vw(16);
-    height: sp-vw(16);
-    filter: brightness(0) invert(1);
-  }
-}
-
-/* ==========================================================================
-   3. 新着求人セクション
-   ========================================================================== */
-
-.new-job {
-  padding: sp-vw(10) 0 sp-vw(60);
-  background-color: $color-light-blue;
-  @include mixins.flex-column-center;
-
-  &-sub-title {
-    margin: sp-vw(35) 0;
-    font-size: sp-vw(25);
-    font-weight: bold;
-  }
-
-  &-content {
-    display: flex;
-    gap: sp-vw(16);
-    width: sp-vw($design-sp-width);
-    margin-bottom: sp-vw(30);
-    overflow-x: auto;
-    -webkit-overflow-scrolling: touch;
-  }
-
-  .content {
-    display: flex;
-    flex-shrink: 0;
-    flex-direction: column;
-    width: sp-vw($inner-sp-max-width);
-    border-radius: $radius-sm;
-    box-shadow: 0 4px 12px rgb(0 0 0 / 4%);
-  }
-
-  .job-img-wrap {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .job-img {
-    display: block;
-    width: sp-vw(470);
-    height: sp-vw(210);
-    object-fit: cover;
-    border-radius: $radius-sm;
-  }
-
-  .job-info {
-    display: flex;
-    flex-direction: column;
-    min-height: sp-vw(100);
-  }
-
-  .job-card-title {
-    margin: sp-vw(5) auto;
-    font-size: sp-vw(16);
-    font-weight: bold;
-    line-height: 1.5;
-    color: $color-black;
-    text-align: left;
-  }
-
-  .job-meta {
-    margin-top: auto;
-  }
-
-  .meta-row {
-    display: flex;
-    flex-direction: row;
-
-    align-items: center;
-  }
-
-  @each $label in $job-labels {
-    .label-#{$label} {
-      min-width: sp-vw(60);
-      padding: sp-vw(5) 0;
-      margin: sp-vw(1);
-      font-size: sp-vw(14);
-      color: $color-black;
-      text-align: center;
-      background-color: $color-light-pink;
-    }
-  }
-
-  .meta-text {
-    margin: 0 sp-vw(15);
-    font-size: sp-vw(15);
-    color: $color-black;
-  }
-
-  &-button {
-    display: block;
-    padding: sp-vw(20) sp-vw(80);
-    font-size: sp-vw(17);
-    box-shadow: $shadow-pink;
-    @include mixins.pink-btn;
-  }
-
-  .slider-wrapper {
-    position: relative;
-    width: sp-vw($inner-sp-max-width);
-    margin-bottom: sp-vw(11);
-
-    @each $dir, $side in (prev: left, next: right) {
-      .#{$dir}-button {
-        #{$side}: sp-vw(0);
-        top: sp-vw(155);
-        width: sp-vw(45);
-        height: sp-vw(45);
-        @include mixins.slider-button;
-      }
-    }
-  }
-}
-
-/* ==========================================================================
-   4. バナー
-   ========================================================================== */
-
-.middle-banner {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: sp-vw(75);
-  background-image:
-    linear-gradient(rgb(0 0 0 / 55%), rgb(0 0 0 / 55%)), url('/icon/banner.jpg');
-  background-position: center 65%;
-  background-size: cover;
-
-  .banner-title {
-    text-align: center;
-  }
-
-  .banner-title,
-  .banner-sub {
-    font-size: sp-vw(13);
-    color: $color-white;
-  }
-
-  .banner-inner {
-    display: flex;
-    flex-direction: column;
-    gap: sp-vw(8);
-    align-items: center;
-  }
-}
