+++ src/app/page.jsx
@@ -1,11 +1,17 @@
 import { Header } from '../components/layout/Header'
-import { Keyvisual } from './_components/Keyvisual'
+import { HomePage } from './_components/HomePage'
+import { SearchForm } from './_components/SearchForm'
+import { NewJob } from './_components/NewJob'
+import { Features } from './_components/Features'
 
 const Top = () => {
   return (
     <>
       <Header />
-      <Keyvisual />
+      <HomePage />
+      <SearchForm />
+      <NewJob />
+      <Features />
     </>
   )
 }
