+++ src/styles/_variables.scss
@@ -9,6 +9,7 @@ $breakpoint-pc: 768;
 // INFO: Color variables
 $color-white: #fff;
 $color-black: #000;
+$color-homepage: #98b1be;
 $color-dark-blue: #deeef9;
 $color-light-blue: #eff6fc;
 $color-light-green: #2eb6aa;
